+++ src/main/java/io/github/astrapi69/time/convert/ZonedDateTimeExtensions.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -25,13 +25,14 @@
 package io.github.astrapi69.time.convert;
 
 import java.time.LocalDateTime;
+import java.time.OffsetDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Date;
 import java.util.Objects;
 
 /**
- * The class {@link ZonedDateTimeExtensions} provides methods for convert {@link LocalDateTime}
+ * The class {@link ZonedDateTimeExtensions} provides methods for convert {@link ZonedDateTime}
  * objects
  */
 public class ZonedDateTimeExtensions
@@ -65,4 +66,34 @@ public class ZonedDateTimeExtensions
 		Objects.requireNonNull(zoneId);
 		return ZonedDateTime.of(localDateTime, zoneId);
 	}
+
+	/**
+	 * Converts the given {@link Date} with the given {@link ZoneId} to a {@link ZonedDateTime}
+	 * object
+	 *
+	 * @param date
+	 *            the date
+	 * @param zone
+	 *            the zone
+	 * @return The {@link ZonedDateTime} object
+	 */
+	public static ZonedDateTime toZonedDateTime(Date date, ZoneId zone)
+	{
+		return OffsetDateTimeExtensions
+			.toZonedDateTime(OffsetDateTimeExtensions.toOffsetDateTime(date, zone));
+	}
+
+	/**
+	 * Converts the given {@link ZonedDateTime} to a {@link OffsetDateTime} object
+	 *
+	 * @param zonedDateTime
+	 *            the {@link ZonedDateTime} object
+	 * @return The {@link OffsetDateTime} object
+	 */
+	public static OffsetDateTime toOffsetDateTime(ZonedDateTime zonedDateTime)
+	{
+		Objects.requireNonNull(zonedDateTime);
+		return zonedDateTime.toOffsetDateTime();
+	}
+
 }