+++ src/test/java/io/github/astrapi69/time/convert/LocalDateTimeExtensionsTest.java
@@ -22,46 +22,67 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.time.stamp;
+package io.github.astrapi69.time.convert;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.sql.Timestamp;
-import java.text.ParseException;
-import java.time.LocalDate;
+import java.time.LocalDateTime;
+import java.time.ZoneId;
+import java.util.Date;
 
+import io.github.astrapi69.time.convert.LocalDateTimeExtensions;
 import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
 
-public class ConvertTimestampExtensionsTest
+import de.alpharogroup.date.CreateDateExtensions;
+
+/**
+ * The unit test class for the class {@link LocalDateTimeExtensions}
+ */
+public class LocalDateTimeExtensionsTest
 {
 
 
 	/**
-	 * Test method for {@link ConvertTimestampExtensions#toTimestamp(LocalDate)}
-	 *
-	 * @throws ParseException
-	 *             occurs when their are problems with parsing the String to Date.
+	 * Test method for {@link LocalDateTimeExtensions#toLocalDateTime(Date, ZoneId)}
 	 */
 	@Test
-	public void testToTimestamp()
+	void toLocalDateTime()
 	{
-		Timestamp actual;
-		Timestamp expected;
-		LocalDate test;
-		test = LocalDate.of(1900, 10, 1);
-		actual = ConvertTimestampExtensions.toTimestamp(test);
-		expected = Timestamp.valueOf(test.atStartOfDay());
-		assertEquals(actual, expected);
+		LocalDateTime actual;
+		LocalDateTime expected;
+		Date date;
+
+		// new scenario
+		date = CreateDateExtensions.newDate(2000, 9, 1);
+		actual = LocalDateTimeExtensions.toLocalDateTime(date, ZoneId.systemDefault());
+		expected = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
+		assertEquals(expected, actual);
 	}
 
 	/**
-	 * Test method for {@link ConvertTimestampExtensions}
+	 * Test method for {@link LocalDateTimeExtensions#toDate(LocalDateTime)}
+	 */
+	@Test
+	void toDate()
+	{
+		Date actual;
+		Date expected;
+		LocalDateTime input;
+
+		input = LocalDateTime.of(2000, 9, 1, 0, 0, 0);
+		actual = LocalDateTimeExtensions.toDate(input);
+		expected = CreateDateExtensions.newDate(2000, 9, 1);
+		assertEquals(expected, actual);
+	}
+
+	/**
+	 * Test method for {@link LocalDateTimeExtensions}
 	 */
 	@Test
 	public void testWithBeanTester()
 	{
 		final BeanTester beanTester = new BeanTester();
-		beanTester.testBean(ConvertTimestampExtensions.class);
+		beanTester.testBean(LocalDateTimeExtensions.class);
 	}
 }